Modifications to the parsing logic.
We aren't necessarily memcached testing suites and it expects output related to that.
As we are parsing one test from a different log file at a time, then we just update the regexes accordingly.

We also read the results twice now: Once for collecting stats and another for timers. This is more expensive, but we are just parsing a log file and it already completes fast enough for a considerably long log (1h+ test run).